Add a method to print optimisation results Move howdy to _util.py
